+++ smarter/smarter/apps/account/signals.py
@@ -27,18 +27,6 @@ Example::
     new_charge_created.send(sender=self.__class__, charge=self)
 """
 
-broker_ready = Signal()
-"""
-Signal sent when a broker achieves a ready state.
-
-Arguments:
-    broker: The broker instance that is ready.
-
-Example::
-
-    broker_ready.send(sender=self.__class__, broker=self)
-"""
-
 cache_invalidate = Signal()
 """
 Signal sent to trigger cache invalidation.